NetBSD: Fix `uucred.cr_ngroups` from `int` to `short`
Link: https://github.com/NetBSD/src/blob/6017cb90fd7d83ed6e45d93129b12d525978c2fa/sys/sys/ucred.h#L52

+++ src/unix/bsd/netbsdlike/netbsd/mod.rs
@@ -394,7 +394,7 @@ s! {
         pub cr_unused: c_ushort,
         pub cr_uid: crate::uid_t,
         pub cr_gid: crate::gid_t,
-        pub cr_ngroups: c_int,
+        pub cr_ngroups: c_short,
         pub cr_groups: [crate::gid_t; NGROUPS_MAX as usize],
     }
